[clang] warnings cleanup

+++ ntoskrnl/mm/section.c
@@ -586,7 +586,7 @@ l_ReadHeaderFromFile:
 
         if(RTL_CONTAINS_FIELD(pioh64OptHeader, cbOptHeaderSize, SizeOfImage))
         {
-            if(pioh64OptHeader->SizeOfImage > MAXULONG_PTR)
+            if ((ULONGLONG)pioh64OptHeader->SizeOfImage > (ULONGLONG)MAXULONG_PTR)
                 DIE(("SizeOfImage exceeds the address space\n"));
 
             ImageSectionObject->ImageInformation.ImageFileSize = pioh64OptHeader->SizeOfImage;
